fix(uv): disable package building for uv to prevent unnecessary build steps
The `package = false` setting is added to the `[tool.uv]` section to explicitly disable the building of the current project as a package. This is done because this configuration file is part of a template setup where the project itself is not intended to be installed as a package, thus speeding up dependency resolution and installation.